Added the clear tasks command that allows the user to clear all the tasks.
Added the edit Task command that allows the user to change the description of the tasks.

Added the due Date to the tasks so the user can put a date limit to complete the tasks.

Changed the todo command so the list of tasks are shown in an embed messsage instead of plain text.